+++ familyhelper-webapi-node/src/main/java/com/dwarfeng/familyhelper/webapi/node/controller/v1/finance/RemindDriverSupportController.java
@@ -1,0 +1,102 @@
+package com.dwarfeng.familyhelper.webapi.node.controller.v1.finance;
+
+import com.dwarfeng.familyhelper.finance.sdk.bean.entity.FastJsonRemindDriverSupport;
+import com.dwarfeng.familyhelper.finance.stack.bean.entity.RemindDriverSupport;
+import com.dwarfeng.familyhelper.webapi.stack.service.finance.RemindDriverSupportResponseService;
+import com.dwarfeng.subgrade.sdk.bean.dto.FastJsonResponseData;
+import com.dwarfeng.subgrade.sdk.bean.dto.JSFixedFastJsonPagedData;
+import com.dwarfeng.subgrade.sdk.bean.dto.PagingUtil;
+import com.dwarfeng.subgrade.sdk.bean.dto.ResponseDataUtil;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.SkipRecord;
+import com.dwarfeng.subgrade.sdk.interceptor.login.LoginRequired;
+import com.dwarfeng.subgrade.stack.bean.BeanTransformer;
+import com.dwarfeng.subgrade.stack.bean.dto.PagedData;
+import com.dwarfeng.subgrade.stack.bean.dto.PagingInfo;
+import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
+import com.dwarfeng.subgrade.stack.exception.ServiceExceptionMapper;
+import org.springframework.web.bind.annotation.*;
+
+import javax.servlet.http.HttpServletRequest;
+
+/**
+ * 提醒驱动器支持控制器。
+ *
+ * @author DwArFeng
+ * @since 1.0.7
+ */
+@RestController("financeRemindDriverSupportController")
+@RequestMapping("/api/v1/finance")
+public class RemindDriverSupportController {
+
+    private final RemindDriverSupportResponseService service;
+    private final ServiceExceptionMapper sem;
+
+    private final BeanTransformer<RemindDriverSupport, FastJsonRemindDriverSupport> beanTransformer;
+
+    public RemindDriverSupportController(
+            RemindDriverSupportResponseService service, ServiceExceptionMapper sem,
+            BeanTransformer<RemindDriverSupport, FastJsonRemindDriverSupport> beanTransformer
+    ) {
+        this.service = service;
+        this.sem = sem;
+        this.beanTransformer = beanTransformer;
+    }
+
+    @GetMapping("/remind-driver-support/{id}/exists")
+    @BehaviorAnalyse
+    @LoginRequired
+    public FastJsonResponseData<Boolean> exists(HttpServletRequest request, @PathVariable("id") String id) {
+        try {
+            boolean exists = service.exists(new StringIdKey(id));
+            return FastJsonResponseData.of(ResponseDataUtil.good(exists));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(e, sem));
+        }
+    }
+
+    @GetMapping("/remind-driver-support/{id}")
+    @BehaviorAnalyse
+    @LoginRequired
+    public FastJsonResponseData<FastJsonRemindDriverSupport> get(HttpServletRequest request, @PathVariable("id") String id) {
+        try {
+            RemindDriverSupport remindDriverSupport = service.get(new StringIdKey(id));
+            return FastJsonResponseData.of(ResponseDataUtil.good(FastJsonRemindDriverSupport.of(remindDriverSupport)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(e, sem));
+        }
+    }
+
+    @GetMapping("/remind-driver-support/all")
+    @BehaviorAnalyse
+    @SkipRecord
+    @LoginRequired
+    public FastJsonResponseData<JSFixedFastJsonPagedData<FastJsonRemindDriverSupport>> all(
+            HttpServletRequest request, @RequestParam("page") int page, @RequestParam("rows") int rows
+    ) {
+        try {
+            PagedData<RemindDriverSupport> all = service.all(new PagingInfo(page, rows));
+            PagedData<FastJsonRemindDriverSupport> transform = PagingUtil.transform(all, beanTransformer);
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonPagedData.of(transform)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(e, sem));
+        }
+    }
+
+    @GetMapping("/remind-driver-support/id-like")
+    @BehaviorAnalyse
+    @SkipRecord
+    @LoginRequired
+    public FastJsonResponseData<JSFixedFastJsonPagedData<FastJsonRemindDriverSupport>> idLike(
+            HttpServletRequest request,
+            @RequestParam("pattern") String pattern, @RequestParam("page") int page, @RequestParam("rows") int rows
+    ) {
+        try {
+            PagedData<RemindDriverSupport> all = service.idLike(pattern, new PagingInfo(page, rows));
+            PagedData<FastJsonRemindDriverSupport> transform = PagingUtil.transform(all, beanTransformer);
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonPagedData.of(transform)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(e, sem));
+        }
+    }
+}